Move methods & children stub into own files
